Fix undefined errors
When running the Q test suite all specs would fail complaining about
`inject` being undefined.

Although there was a function wrapping said function it still was not
defined and would cause a reference error well before the actual
`isDefined()` function was called.

Considering the use of global variables the *is defined* lgic needs to
be in the place of use and not proxied by a function.

Replace `isDefined()` with: (browser compatible syntax from
CoffeeScript)

| Type     | How to check                                           |
|==========|========================================================|
| globals  | `typeof variable !== 'undefined' && variable !== null` |
| function | `function(x) { return x != null; }`                    |
| property | `var x = {}; return !x.missing;`                       |