Pass the correct context to the runs function
The jasmine runs function passes the context to the current Spec to it.
This allows you to attach properties to this and have it available
inside the runs function. Because of this variable you want to define in
the beforeEach are accessible.

Add specs around this to be sure that the context is set.